Wrap code in an interval, check for active mail_body
This fixes #17. On the way to fixing it, I notice 
a bug with flipping between mails using < and > or
last and next page button/shortcut.e We need to find 
link AFTER the next content is added to the DOM, 
check for the right mail_body, and then query the
links within that mail_body; otherwise, the links
from the previous mail_body will get appended.