Open tabs next to the current one
When clicking or cmd-clicking a link in a tab, Chrome opens a new tab
next to the current one, or, if many have been clicked, at the end of
that "local stack". `chrome.tabs.create` opens new tabs in the last
position in the window by default, so update the code to mirror the
mouse-click behavior.