Fix links so they are clickable
It appears as though the `ar7` is now associated with another `ask`
class in Gmail now.

This leads to `pointer-events: none` being added to the `ar7` class
making the links in Gmail no longer clickable.

Since the ar7 class is still used for color, font, and alignment,
override the pointer-events as auto in the link style.